feat: 设置接口访问权限
