+++ middleWare/UpdateCookie.js
@@ -4,13 +4,21 @@
  */
 
 import authorization from "../conf/authorization.js";
+import { checkUrl, getRouterName } from "../util/urlUtil.js";
 
 /**
  * @description 每次用户发送请求后更新cookie，修改cookie过期时间
  */
 export function updateCookie(req, res, next) {
+    const url = checkUrl(req.url);
+    const routerName = getRouterName(url);
+
     // 放行匿名请求
-    if (authorization["anonymous"].hasOwnProperty(req.url)) next();
+    if (
+        authorization["anonymous"].hasOwnProperty(url) ||
+        authorization["anonymous"].hasOwnProperty(`/${routerName}/*`)
+    )
+        next();
     else {
         // 更新cookie
         res.cookie("name", req.signedCookies.name, {
